media_management_api/requirements/base.txt
Django==1.11.12
psycopg2==2.6.1
redis==2.10.3
django-redis-cache==1.7.1
hiredis==0.2.0
djangorestframework==3.8.2
markdown==2.6.11
mock==1.3.0
django-filter==1.1.0
django-cors-headers==2.2.0
boto==2.48.0
pillow==5.1.0
python-magic==0.4.15
urllib3[secure]==1.22
requests==2.18.4